+++ pkgs/cli/src/commands/hats.ts
@@ -7,7 +7,6 @@ import {
 	mintHat,
 } from "../modules/hatsProtocol";
 import { PinataSDK } from "pinata-web3";
-import { getJwt, setJwt } from "../services/hats";
 import { getAccount } from "../services/wallet";
 import { publicClient, rootProgram, walletClient } from "..";
 import { Address } from "viem";
@@ -71,96 +70,6 @@ hatsCommands
 		const wearer = await getWearerInfo(address, rootProgram.opts().chain);
 
 		console.log(wearer);
-	});
-
-/**
- * PinataのJWTを設定するコマンド
- */
-hatsCommands
-	.command("pinata")
-	.description("Set a jwt of Pinata")
-	.requiredOption("--jwt <JWT>")
-	.action(({ jwt }) => {
-		setJwt(jwt);
-	});
-
-/**
- * Hatのメタデータをipfs上にアップロードするコマンド
- */
-interface Responsibility {
-	label: string;
-	description?: string;
-	link?: string;
-}
-interface Eligibility {
-	manual: boolean;
-	criteria: string[];
-}
-interface Toggle {
-	manual: boolean;
-	criteria: string[];
-}
-hatsCommands
-	.command("upload")
-	.description("Upload the hat metadata on ipfs.")
-	.requiredOption("-n, --name <name>", "Hat Name")
-	.option("-d, --description <description>", "Hat Details")
-	.option(
-		"-r, --responsibility <label>,<description>,<link>",
-		"Responsibility (may be specified multiple times to define multiple responsibilities)",
-		(value, previous: Responsibility[]) => {
-			const [label, description, link] = value.split(",");
-			return previous ? previous.concat([{ label, description, link }]) : [{ label, description, link }];
-		},
-		[]
-	)
-	.option(
-		"-a, --authority <authority>",
-		"Authority (may be specified multiple times to define multiple authorities)",
-		(value, previous: string[]) => previous ? previous.concat([value]) : [value],
-		[]
-	)
-	.option(
-		"-e, --eligibility <manual>,<criteria...>",
-		"Eligibility (<manual> is a boolean value, <criteria... > can be specified multiple times, separated by commas, to define multiple criteria.)",
-		(value) => {
-			const [manual, ...criteria] = value.split(",");
-			return { manual: manual === "true", criteria } satisfies Eligibility;
-		}
-	)
-	.option(
-		"-t, --toggle <manual> <criteria...>",
-		"Toggle (<manual> is a boolean value, <criteria... > can be specified multiple times, separated by spaces, to define multiple criteria.)",
-		(value) => {
-			const [manual, ...criteria] = value.split(",");
-			return { manual: manual === "true", criteria } satisfies Toggle;
-		}
-	)
-	.action(async ({
-		name,
-		description,
-		responsibility,
-		authority,
-		eligibility,
-		toggle,
-	}) => {
-		const { jwt } = getJwt();
-	
-		const pinata = new PinataSDK({ pinataJwt: jwt });
-
-		const upload = await pinata.upload.json({
-			"type": "1.0",
-			"data": {
-				name,
-				description,
-				responsibilities: responsibility,
-				authorities: authority,
-				eligibility,
-				toggle
-			}
-		});
-
-		console.log("CID:", upload.IpfsHash);
 	});
 
 /**
