+++ pkgs/cli/src/index.ts
@@ -8,6 +8,7 @@ import { getPublicClient } from "./modules/viem";
 import { getWalletClient } from "./services/wallet";
 import { skipPreActionCommands } from "./config";
 import { bigbangCommands } from "./commands/bigbang";
+import { pinataCommands } from "./commands/pinata";
 
 export const rootProgram = new Command();
 
@@ -33,5 +34,6 @@ rootProgram
 rootProgram.addCommand(bigbangCommands);
 rootProgram.addCommand(hatsCommands);
 rootProgram.addCommand(walletCommands);
+rootProgram.addCommand(pinataCommands);
 
 rootProgram.parse(process.argv);
