+++ pkgs/cli/src/services/pinata.ts
